+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/cores/structure/DriveOp.java
@@ -24,6 +24,18 @@ public strictfp class DriveOp implements Interfaces.HardwareController, Interfac
 		return instance;
 	}
 
+	@NonNull
+	@Contract("_, _, _ -> new")
+	public static Action build(final double x, final double y, final double turn) {
+		return new SimpleDriveAction(x, y, turn);
+	}
+
+	@NonNull
+	@Contract("_, _, _, _ -> new")
+	public static Action build(final double x, final double y, final double turn, final double bufPower) {
+		return new SimpleDriveAction(x * bufPower, y * bufPower, turn * bufPower);
+	}
+
 	@Override
 	public void connect() {
 		chassisCtrl = new ChassisCtrl(HardwareDatabase.leftFront, HardwareDatabase.leftRear, HardwareDatabase.rightFront, HardwareDatabase.rightRear);
